+++ noodles-cram/src/data_container/compression_header/preservation_map/tag_ids_dictionary/builder.rs
@@ -10,8 +10,14 @@ pub struct Builder {
 
 impl Builder {
     pub fn update(&mut self, record: &Record) {
-        let keys: Vec<_> = record.tags().values().map(|field| field.into()).collect();
+        let keys: Vec<_> = record
+            .tags()
+            .iter()
+            .map(|(tag, value)| Key::new(tag, value.ty()))
+            .collect();
+
         let next_index = self.keys_indices.len();
+
         self.keys_indices.entry(keys).or_insert(next_index);
     }