+++ noodles-sam/src/writer/record/data/field.rs
@@ -4,19 +4,19 @@ mod value;
 use std::io::{self, Write};
 
 use self::{tag::write_tag, value::write_value};
-use crate::record::data::Field;
+use crate::record::data::field::{Tag, Value};
 
-pub fn write_field<W>(writer: &mut W, field: &Field) -> io::Result<()>
+pub fn write_field<W>(writer: &mut W, tag: Tag, value: &Value) -> io::Result<()>
 where
     W: Write,
 {
     const DELIMITER: u8 = b':';
 
-    write_tag(writer, field.tag())?;
+    write_tag(writer, tag)?;
     writer.write_all(&[DELIMITER])?;
-    value::write_type(writer, field.value().ty())?;
+    value::write_type(writer, value.ty())?;
     writer.write_all(&[DELIMITER])?;
-    write_value(writer, field.value())?;
+    write_value(writer, value)?;
 
     Ok(())
 }
@@ -27,13 +27,10 @@ mod tests {
 
     #[test]
     fn test_write_field() -> io::Result<()> {
-        use crate::record::data::field::{Tag, Value};
-
         let mut buf = Vec::new();
-        let field = Field::new(Tag::AlignmentHitCount, Value::from(1));
-        write_field(&mut buf, &field)?;
+        let (tag, value) = (Tag::AlignmentHitCount, Value::from(1));
+        write_field(&mut buf, tag, &value)?;
         assert_eq!(buf, b"NH:i:1");
-
         Ok(())
     }
 }